fix: residence is where you live, not what your passport says
Three from the review of ed6fcd33.

The residence hint told people to pick the country their documents show. A
resident of Argentina holding a Venezuelan passport would pick Venezuela — and
eligibility for the card and the local rails keys off KYC residence, not the
document's issuing country, with foreign passports accepted. The wrong pick
changes the restrictions and rail preview stored on the account. The hint now
describes where the person legally lives and has their address, and keeps the
travelling clarification that made the original copy useful.

LockCardModal's unlock variant passed a fragment whose children are all absent
until an error occurs. ActionModal reads a truthy  as 'this modal has
a body', so that bought the head's 12px, an empty body wrapper and the ctas'
24px — a 36px head-to-button gap on a modal the component's own contract says
should be 24px. The caller now passes undefined when nothing renders. The
gallery shoots the lock variant, so the capture never covered this.

The dev surfaces route gated correctly but still imported the registry
statically, pulling it and its fifty-odd component imports into the route graph
of every production and native-export build. It now loads behind
DEV_TOOLS_ENABLED through a dynamic import, the shape /dev/fixtures documents.
