+++ src/components/Card/LockCardModal.tsx
@@ -121,6 +121,22 @@ const LockCardModal: FC<Props> = ({ cardId, mode, isOpen, onClose }) => {
 
     const isSuccess = phase === 'success'
 
+    const showError = phase === 'error' && !!error
+    const showSlide = mode === 'lock'
+    const hasBody = !isSuccess && (showError || showSlide)
+    const bodyContent = (
+        <>
+            {showError && <p className="text-body-s text-foreground-error">{error}</p>}
+            {showSlide && (
+                <SlideToConfirm
+                    label={phase === 'loading' ? t('lockModal.locking') : t('lockModal.slideToLock')}
+                    onConfirm={run}
+                    disabled={phase === 'loading'}
+                />
+            )}
+        </>
+    )
+
     return (
         <ActionModal
             visible={isOpen}
@@ -131,20 +147,11 @@ const LockCardModal: FC<Props> = ({ cardId, mode, isOpen, onClose }) => {
             icon="lock"
             title={t(isSuccess ? copyKeys.success : copyKeys.title)}
             description={t(isSuccess ? copyKeys.successBody : copyKeys.body)}
-            content={
-                isSuccess ? undefined : (
-                    <>
-                        {phase === 'error' && error && <p className="text-body-s text-foreground-error">{error}</p>}
-                        {mode === 'lock' && (
-                            <SlideToConfirm
-                                label={phase === 'loading' ? t('lockModal.locking') : t('lockModal.slideToLock')}
-                                onConfirm={run}
-                                disabled={phase === 'loading'}
-                            />
-                        )}
-                    </>
-                )
-            }
+            /* undefined, not an empty fragment: ActionModal reads a truthy
+               `content` as "this modal has a body" and spaces it accordingly, so
+               a fragment whose children are all absent buys the head margin and
+               an empty wrapper for nothing. */
+            content={hasBody ? bodyContent : undefined}
             ctas={
                 !isSuccess && mode === 'unlock'
                     ? [
